chore: fixing checkPermissions

+++ admin/adminControllers/adminControllers.js
@@ -1,4 +1,5 @@
 const adminServices = require("../adminServices/adminServices");
+const adminAuthServices = require("../adminServices/adminAuthServices");
 
 //Admin CRUD
 
@@ -32,20 +33,30 @@ const readAllAdmins = async (req, res) => {
 
 const updateAdmin = async (req, res) => {
   try {
-    const adminData = {
-      firstname: req.body.firstname,
-      lastname: req.body.lastname,
-      phone: req.body.phone,
-      area: req.body.area,
-      username: req.body.username,
-      email: req.body.email,
-    };
-
-    const updatedAdmin = await adminServices.updateAdmin(
-      req.params.id,
-      adminData
+    let checkPermissions = await adminAuthServices.checkPermissions(
+      req.body.adminToken
     );
-    return res.send(updatedAdmin);
+    if (checkPermissions.modifyAdmins) {
+      const adminData = {
+        firstname: req.body.firstname,
+        lastname: req.body.lastname,
+        phone: req.body.phone,
+        area: req.body.area,
+        username: req.body.username,
+        email: req.body.email,
+      };
+
+      const updatedAdmin = await adminServices.updateAdmin(
+        req.params.id,
+        adminData
+      );
+      return res.send(updatedAdmin);
+    } else {
+      return res.send({
+        success: false,
+        message: "No tienes autorización para realizar esta acción.",
+      });
+    }
   } catch (err) {
     res.status(500).send(err);
   }
@@ -53,8 +64,18 @@ const updateAdmin = async (req, res) => {
 
 const deleteAdmin = async (req, res) => {
   try {
-    const deletedAdmin = await adminServices.deleteAdmin(req.params.username);
-    res.send(deletedAdmin);
+    let checkPermissions = await adminAuthServices.checkPermissions(
+      req.body.adminToken
+    );
+    if (checkPermissions.modifyAdmins) {
+      const deletedAdmin = await adminServices.deleteAdmin(req.params.username);
+      res.send(deletedAdmin);
+    } else {
+      return res.send({
+        success: false,
+        message: "No tienes autorización para realizar esta acción.",
+      });
+    }
   } catch (err) {
     res.status(500).send(err);
   }
@@ -63,8 +84,18 @@ const deleteAdmin = async (req, res) => {
 // Admin Role CRUD
 const createAdminRole = async (req, res) => {
   try {
-    const createdAdminRole = await adminServices.createAdminRole(req.body);
-    res.send(createdAdminRole);
+    let checkPermissions = await adminAuthServices.checkPermissions(
+      req.body.adminToken
+    );
+    if (checkPermissions.modifyAdmins) {
+      const createdAdminRole = await adminServices.createAdminRole(req.body);
+      res.send(createdAdminRole);
+    } else {
+      return res.send({
+        success: false,
+        message: "No tienes autorización para realizar esta acción.",
+      });
+    }
   } catch (err) {
     res.status(500).send(err);
   }
@@ -81,44 +112,66 @@ const readAdminRoles = async (req, res) => {
 
 const updateAdminRole = async (req, res) => {
   try {
-    const idToUpdate = req.params.id;
-    const updateRole = {
-      area: req.body.area,
-      detailOrder: req.body.detailOrder,
-      detailPay: req.body.detailPay,
-      orderStatus: req.body.orderStatus,
-      createOrder: req.body.createOrder,
-      createProduct: req.body.createProduct,
-      deleteProduct: req.body.deleteProduct,
-      createDiscount: req.body.createDiscount,
-      deleteDiscount: req.body.deleteDiscount,
-      modifyBanners: req.body.modifyBanners,
-      modifyTermsAndCo: req.body.modifyTermsAndCo,
-      createPaymentMethod: req.body.createPaymentMethod,
-      deletePaymentMethod: req.body.deletePaymentMethod,
-      createShippingMethod: req.body.createShippingMethod,
-      deleteShippingMethod: req.body.deleteShippingMethod,
-      modifyDollar: req.body.modifyDollar,
-      prixerBan: req.body.prixerBan,
-      createTestimonial: req.body.createTestimonial,
-      deleteTestimonial: req.body.deleteTestimonial,
-      modifyAdmins: req.body.modifyAdmins,
-      setPrixerBalance: req.body.setPrixerBalance,
-      readMovements: req.body.readMovements,
-    };
-    const updatedAdminRole = await adminServices.updateAdminRole(
-      idToUpdate,
-      updateRole
+    let checkPermissions = await adminAuthServices.checkPermissions(
+      req.body.adminToken
     );
-    return res.send(updatedAdminRole);
+    if (checkPermissions.modifyAdmins) {
+      const idToUpdate = req.params.id;
+      const updateRole = {
+        area: req.body.area,
+        detailOrder: req.body.detailOrder,
+        detailPay: req.body.detailPay,
+        orderStatus: req.body.orderStatus,
+        createOrder: req.body.createOrder,
+        createProduct: req.body.createProduct,
+        deleteProduct: req.body.deleteProduct,
+        createDiscount: req.body.createDiscount,
+        deleteDiscount: req.body.deleteDiscount,
+        modifyBanners: req.body.modifyBanners,
+        modifyTermsAndCo: req.body.modifyTermsAndCo,
+        createPaymentMethod: req.body.createPaymentMethod,
+        deletePaymentMethod: req.body.deletePaymentMethod,
+        createShippingMethod: req.body.createShippingMethod,
+        deleteShippingMethod: req.body.deleteShippingMethod,
+        modifyDollar: req.body.modifyDollar,
+        prixerBan: req.body.prixerBan,
+        createTestimonial: req.body.createTestimonial,
+        deleteTestimonial: req.body.deleteTestimonial,
+        modifyAdmins: req.body.modifyAdmins,
+        setPrixerBalance: req.body.setPrixerBalance,
+        readMovements: req.body.readMovements,
+      };
+      const updatedAdminRole = await adminServices.updateAdminRole(
+        idToUpdate,
+        updateRole
+      );
+      return res.send(updatedAdminRole);
+    } else {
+      return res.send({
+        success: false,
+        message: "No tienes autorización para realizar esta acción.",
+      });
+    }
   } catch (err) {
     res.status(500).send(err);
   }
 };
 const deleteAdminRole = async (req, res) => {
   try {
-    const deletedAdminRole = await adminServices.deleteAdminRole(req.params.id);
-    res.send(deletedAdminRole);
+    let checkPermissions = await adminAuthServices.checkPermissions(
+      req.body.adminToken
+    );
+    if (checkPermissions.modifyAdmins) {
+      const deletedAdminRole = await adminServices.deleteAdminRole(
+        req.params.id
+      );
+      res.send(deletedAdminRole);
+    } else {
+      return res.send({
+        success: false,
+        message: "No tienes autorización para realizar esta acción.",
+      });
+    }
   } catch (err) {
     res.status(500).send(err);
   }
